backend/requirements.txt
# FastAPI and server
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20

# AI/ML Services
faster-whisper==1.1.1
kokoro-onnx>=0.4.0
soundfile>=0.13.0
onnxruntime>=1.19.0

# Video/Audio Processing
ffmpeg-python==0.2.0

# Utilities
pydantic==2.10.4
python-dotenv==1.0.1
aiofiles==24.1.0
